requirements.txt
quart
python-telegram-bot
python-dotenv
openai
uvicorn